refactor(transcoding): streamline transcoding logic by consolidating stream parameter handling and enhancing alias mapping
